Refactor folder placement
